requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
cx_Freeze>=6.15.0
